Fix TTY for GPG usage in sandbox
Signing of GPG packages isn't working in a sandbox because of some TTY issues if you ssh to the zone and try to build / sign packages in the sandbox. Which result in the following error:

	gpg: problem with the agent: Inappropriate ioctl for device
	gpg: signing failed: Operation cancelled
	pkg_admin: GPG could not create signature

The problem could be solved by export the `GPG_TTY` variable in the sandbox as well.