Correctly set a soft ulimit for cputime.
It's a bit embarrassing how long this has been wrong, pkgsrc supports
resetting ulimits back to the hard limits on a per-package basis, but as
I was setting both soft and hard limits it had no effect.

Reduce every build back to the default of 1 hour so we can identify
packages that fail and fix legitimate requirements.